+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 Changelog
 =========
 
+Version 0.2.3.0
+---------------
+
+*November 21, 2019*
+
+<https://github.com/mstksg/advent-of-code-api/releases/tag/v0.2.3.0>
+
+*   Add API commands for daily and global leaderboards
+*   Move some of the data types to be in their own module
+
 Version 0.2.2.1
 ---------------